cicd: build and release workflow (#126)
* cicd: update build script to include archiving

* cicd: build and release workflow

+++ scripts/build.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 bin=burnit
 module_path=github.com/RedeployAB/$bin
-os=linux
-arch=amd64
-build_root=build/
-bin_path=$build_root/$bin
+build_root=build
+platform=linux/amd64
+
 version=""
-container=0
+image=0
+archive=0
+cleanup=1
+skip_tests=0
 
 for arg in "$@"
 do
@@ -16,22 +18,31 @@ do
       version=$1
       shift
       ;;
-    --os)
+    --platform)
       shift
-      os=$1
-      shift
-      ;;
-    --arch)
-      shift
-      arch=$1
+      platform=$1
       shift
       ;;
     --image)
-      container=1
+      image=1
+      shift
+      ;;
+    --archive)
+      archive=1
+      shift
+      ;;
+    --no-cleanup)
+      cleanup=0
+      shift
+      ;;
+    --skip-tests)
+      skip_tests=1
       shift
       ;;
   esac
 done
+
+cwd=$(pwd)
 
 if [ -z $bin ]; then
   bin=$(echo $(basename $(pwd)))
@@ -46,22 +57,42 @@ if [ -z $build_root ]; then
   exit 1
 fi
 
-if [ -d $build_root ]; then
+if [ $cleanup -eq 1 ] && [ -d $build_root ]; then
   rm -rf $build_root/*
 fi
 
 mkdir -p $build_root
 
-go test ./...
-if [ $? -ne 0 ]; then
-  exit 1
+if [ $skip_tests -eq 0 ]; then
+  go test ./...
+  if [ $? -ne 0 ]; then
+    exit 1
+  fi
 fi
 
-CGO_ENABLED=0 GOOS=$os GOARCH=$arch go build \
-  -o $bin_path \
+platforms=(${platform//,/ })
+for p in "${platforms[@]}"; do
+  os_arch=(${p//\// })
+  os=${os_arch[0]}
+  arch=${os_arch[1]}
+
+  bin_path=$build_root/$os/$arch
+  CGO_ENABLED=0 GOOS=$os GOARCH=$arch go build \
+  -o $bin_path/$bin \
   -ldflags="-s -w -X '$module_path/internal/version.version=$version'" \
   -trimpath main.go
 
-if [ $container -eq 1 ] && [ "$os" == "linux" ]; then
-  docker build -t $bin:$version --platform $os/$arch .
+  if [ $archive -eq 1 ]; then
+    cd $bin_path
+    targz=$bin-$version-$os-$arch.tar.gz
+    tar -czf $targz *
+    mv $targz $cwd/$build_root/$targz
+    cd $cwd
+  fi
+done
+
+if [ $image -eq 1 ]; then
+  docker buildx create --name multiarch --use --bootstrap
+  docker build -t $bin:$version --platform $platform .
+  docker buildx rm multiarch
 fi
